misha_audio/requirements.txt
pyttsx3>=2.90
pypdf>=4.2.0
python-docx>=1.1.0
